python-client/requirements.txt
pandas==2.1.4
plyer==2.1.0
pycryptodome==3.20.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
